+++ o3research/marketing/__init__.py
@@ -3,6 +3,7 @@
 from .google_ads_agent import GoogleAdsCampaignAgent
 from .meta_ads_agent import MetaAdsAgent
 from .budget_allocator import BudgetAllocatorAgent
+from .budgeter import Budgeter
 from .funnel_planner import FunnelPlannerAgent
 from .creative_prompt_agent import CreativePromptAgent
 from .creative import CreativeAgent
@@ -11,6 +12,7 @@ __all__ = [
     "GoogleAdsCampaignAgent",
     "MetaAdsAgent",
     "BudgetAllocatorAgent",
+    "Budgeter",
     "FunnelPlannerAgent",
     "CreativePromptAgent",
     "CreativeAgent",